requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
boto3==1.26.77
botocore==1.29.77
calculation==8.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.2
colorama==0.4.6
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.4
django-allauth==0.52.0
django-calculation==1.0.0
django-countries==7.5.1
django-crispy-forms==2.0
django-entangled==0.5.3
django-jsonform==2.15.0
django-phonenumber-field==7.0.2
django-storages==1.13.2
Flask==2.1.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.2
paystackapi==2.1.0
phonenumbers==8.13.6
Pillow==9.3.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.6
psycopg2==2.9.5
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pypaystack==1.24
pypaystack2==1.0.3
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-flutterwave==0.7.9
python3-openid==3.2.0
qrcode==7.4.2
rave-python==1.2.16
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.1.1
whitenoise==6.3.0
zipp==3.15.0
